feat: assing space API

+++ src/application/use-cases/config/delete-config.command.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { ConfigHistoryService } from '../../domain/config-history/config-history.service';
 import { ConfigService } from '../../domain/config/config.service';
 
@@ -11,6 +11,12 @@ export class DeleteConfigCommand {
 
   async execute(space: string, name: string, updateReason: string, updatedBy: string): Promise<void> {
     const config = await this.configService.findByName(space, name);
+    if (!config) {
+      throw new NotFoundException(`Config with name ${name} not found`);
+    }
+    if (config.isDisabled) {
+      throw new Error(`Config with name ${name} is already disabled`);
+    }
     await this.configService.deleteByName(space, name);
 
     await this.configHistoryService.createHistory({
